Referral API parse FIX - To parse the Token value

+++ mobile-client-android/sentinel_android/app/src/main/java/sentinelgroup/io/sentinel/network/model/Referral.java
@@ -1,6 +1,6 @@
 package sentinelgroup.io.sentinel.network.model;
 
 public class Referral {
-    public int count;
-    public int amount;
+    public long count;
+    public long amount;
 }
